+++ src/routes/graphql/types/nonNullTypes.ts
@@ -1,6 +1,6 @@
 import { GraphQLNonNull } from "graphql";
 import { UUIDType } from "./uuid.js";
-import { memberEnumType } from "../memberQuery/memberObjectType.js";
+import { memberEnumType } from "../queryTypes/memberQuery/memberObjectType.js";
 
 const nonNullUUIDType = new GraphQLNonNull(UUIDType);
 const nonNullMemberEnumType = new GraphQLNonNull(memberEnumType);
